pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oss-crs"
version = "0.1.0"
description = "Cyber Reasoning System (CRS) build and run tools for bug finding and fixing"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Team Atlanta"}
]
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0.3",
    "Jinja2>=3.1.6",
    "python-dotenv>=1.0.1",
    "docker>=7.1.0",
    "ruff>=0.14.5",
    "pyright>=1.1.407",
    "gitpython>=3.1.45",
    "pydantic>=2.12.5",
    "rich>=13.7.0",
    "questionary>=2.1.1",
]

[project.scripts]
crs-compose = "crs.src.cli.crs_compose:main"
crs = "crs.src.cli.crs:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "watchdog>=4.0.0",
]

[tool.setuptools]
packages = [
    "crs", "crs.src",
    "crs_registry"
]

[tool.setuptools.package-data]
"crs_registry" = ["*/*.yaml"]

[dependency-groups]
dev = [
    "ty>=0.0.4",
]

[tool.ty.src]
include = []

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (require network access, may be slow)",
]
